Move burnham operators to new utils module
